# src/docs/swagger.yaml
openapi: '3.0.0'
info:
  title: Recipe Sharing Community
  description: |
    The Recipe Sharing Community API allows users to share their favorite recipes and search for recipes based on ingredients, title, or description. Users can also perform actions such as adding, updating, deleting recipes, upvoting or downvoting recipes, and adding comments. Additionally, the platform includes features like followers and following functionality.

    Some recipes are marked as premium and require a subscription plan for access. Admins and super admins manage subscription plans, oversee user management, and handle premium content access. While regular users can sign up and log in to the application, super admins have advanced permissions to add, update, or delete both users and recipes.
  version: '1.0.0'
  contact:
    name: Jakirul Islam Hakim
    email: jakirulisamhakim@gmail.com
    url: https://hakim-portfolio-web.vercel.app/
servers:
  - url: http://localhost:3000/api/v1
    description: Local server
...
paths:
...
  /auth/login:
    post:
      operationId: loginUser
      tags:
        - auth
      summary: Login an existing user .
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  description: The username or email of the user.
                  type: string
                  example: hakim_h4
                password:
                  type: string
                  example: 12345
              required: ['identifier', 'password']
      responses:
        200:
          description: Successfully login user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: *successTrue
                  message:
                    type: string
                    example: Login successful. Welcome back!
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6Ikhha2ltIiwiaWF0IjoxNTE2MjM5MDIyfQ._iXTScqTXCc4-prKy6isth5zp2HKquyddAriFO6n80A
                  data:
                    type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/userSchema'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
  /auth/forget-password:
    post:
      operationId: forgetPassword
      tags:
        - auth
      summary: Initiates the process of resetting the password .
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: hakim@gmail.com
      responses:
        200:
          description: Requested forget password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: *successTrue
                  message:
                    type: string
                    example: We’ve sent you an email with instructions to reset your password.
                  data:
                    type: object
                    example:
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /auth/reset-password/{token}:
    post:
      operationId: resetPassword
      tags:
        - auth
      summary: Completes the password reset process.
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: The token received in the email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: newPass123
      responses:
        200:
          description: Password reset successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: *successTrue
                  message:
                    type: string
                    example: Password has been successfully reset.
                  data:
                    type: object
                    properties:
                      payload:
                        type: object
                        example: null
                      links:
                        $ref: '#/components/schemas/linksSchema'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # request body schemas
...
  schemas:
...
    userSchema:
      allOf:
        - type: object
          properties:
            _id: *_id
            fullName:
              type: string
              example: Jakirul Islam Hakim
            username:
              type: string
              example: hakim224242
            email:
              type: string
              format: email
              example: hakim@gmail.com
            profileImg:
              type: string
              format: url
              example: https://i.ibb.co.com/4jr3Rn6/no-images.png
            bio:
              type: string
              nullable: true
              example: ~
            role:
              type: string
              enum:
                - superAdmin
                - admin
                - user
              example: user
            isUserPremium:
              type: boolean
              example: false
            subscriptionEndDate:
              type: string
              nullable: true
              example: null
            # isDeleted:
            #   type: boolean
            #   example: false
            # passwordChangeAt:
            #   type: string
            #   format: date-time
            #   example: 2024-12-27T17:02:42.861Z
          required:
            - _id
            - fullName
            - username
            - email
            - profileImg
            - bio
            - role
            - isUserPremium
            - subscriptionEndDate
            - isDeleted
            - passwordChangeAt
        - $ref: '#/components/schemas/timeStamps'
    # recipe schemas
...
    timeStamps:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
          example: 2024-12-27T17:02:42.861Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-12-27T17:02:42.861Z
      required: ['createdAt', 'updatedAt']
    # links reusable schema
    linksSchema:
      type: object
      properties:
        self:
          type: object
          properties:
            href:
              type: string
              format: url
              example: '/auth/register'
            method:
              type: string
              example: POST
        login:
          type: object
          properties:
            href:
              type: string
              format: url
              example: '/auth/login'
            method:
              type: string
              example: POST
    # pagination schema
...
  responses:
    400:
      description: Bad request Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: &successFalse
                type: boolean
                example: false
              message:
                type: string
                example: Validation error!
              errorSources:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      example: email
                    message:
                      type: string
                      example: email is required
    401:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: *successFalse
              message: &msg401
                type: string
                example: Authentication is required to access this resource.
              errorSources:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      example: ''
                    message: *msg401
    403:
      description: Forbidden Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: *successFalse
              message: &msg403
                type: string
                example: You do not have permission to access this resource.
              errorSources:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      example: ''
                    message: *msg403
    404:
      description: Not Found Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: *successFalse
              message: &msg404
                type: string
                example: Account not found. Please check your credentials or sign up.
              errorSources:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      example: ''
                    message: *msg404
...
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: *successFalse
              message: &msg500
                type: string
                example: An unexpected error occurred. Please try again later.
              errorSources:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      example: ''
                    message: *msg500
  # parameters
...